+++ src/main/java/com/retriage/retriage/models/PatientPoolTmp.java
@@ -17,7 +17,7 @@ public class PatientPoolTmp {
     @Unique
     private String name;
 
-    private int processTime;
+    private Long processTime;
 
     private boolean useable;
